created views directory with file for each view class
