+++ team_production_system/models.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 # from django.db import models
 # from django.contrib.auth.models import AbstractUser
 # from django.conf import settings
