+++ team_production_system/views/availability_delete.py
@@ -1,0 +1,21 @@
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticated
+from django.http import Http404
+from team_production_system.serializers import AvailabilitySerializer
+from team_production_system.models import Availability
+from team_production_system.custom_permissions import IsOwnerOrAdmin
+
+
+class AvailabilityDeleteView(generics.DestroyAPIView):
+    serializer_class = AvailabilitySerializer
+    permission_classes = [IsAuthenticated, IsOwnerOrAdmin]
+
+    def get_object(self):
+        try:
+            # Get the Availability instance for the logged in user
+            availability = Availability.objects.select_related(
+                'mentor__user').get(id=self.kwargs['pk'])
+            self.check_object_permissions(self.request, availability)
+            return availability
+        except Availability.DoesNotExist:
+            raise Http404("No Availability matches the given query.")
