+++ team_production_system/views/session_list.py
@@ -1,0 +1,21 @@
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticated
+from django.utils import timezone
+from django.db.models import Q
+from team_production_system.serializers import SessionSerializer
+from team_production_system.models import Session
+
+
+class SessionView(generics.ListAPIView):
+    queryset = Session.objects.all()
+    serializer_class = SessionSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        # Get sessions for the logged in user
+        # Exclude sessions that have already ended and
+        # order by sessions that are coming up next first.
+        return Session.objects.filter(Q(mentor__user=self.request.user) |
+                                      Q(mentee__user=self.request.user),
+                                      start_time__gt=timezone.now()
+                                      ).order_by('start_time')
